@@ -1,6 +1,6 @@
 [metadata]
 name = unifi-cam-proxy
-version = 0.0.3
+version = 0.1.0
 description = Unifi NVR-compatible camera proxy
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -11,22 +11,42 @@ license = MIT
 license_file = LICENSE
 classifiers =
     License :: OSI Approved :: MIT License
-    Programming Language :: Python :: 2.7
+    Programming Language :: Python :: 3.8
 
 [options]
 packages = find:
 install_requires =
+    aiohttp
+    aiofiles
+    asyncio-mqtt
+    backoff
     coloredlogs
-    requests
-    websocket-client
-    flvlib
     hikvisionapi
+    websockets
     xmltodict
-python_requires = ~=2.7
+python_requires = >3.7.0
 
 [options.entry_points]
 console_scripts =
     unifi-cam-proxy=unifi.main:main
 
+[options.extras_require]
+test =
+    black
+    isort
+    flake8
+    flake8-bugbear
+    pyre-check
+    pytest
+all =
+    %(test)s
+
 [bdist_wheel]
 universal = True
+
+[flake8]
+max-line-length = 80
+select = C,E,F,W,B,B950
+ignore = E203, E501, W503
+exclude = .git,__pycache__,venv,build,dist
+
